Moved custom parser into parsers
