+++ src/index.ts
@@ -19,6 +19,6 @@ export { default as SphericalSegmentGroup } from './geometry/SphericalSegmentGro
 export { default as TorusSegmentGroup } from './geometry/TorusSegmentGroup';
 export { default as TrapeziumGroup } from './geometry/TrapeziumGroup';
 export { Attribute }  from './geometry/PrimitiveGroup';
-export { parseFullCustomFile, parseMultipleCustomFiles } from './customFileParser/main';
+export { parseFullCustomFile, parseMultipleCustomFiles } from './parsers/i3d/main';
 export { computeBoundingBox } from './geometry/GeometryUtils';
 export { TreeIndexNodeIdMap, ColorMap } from './parsers/parseUtils';
