+++ src/parsers/FibonacciDecoder.ts
