+++ src/parsers/i3d/PropertyLoader.ts
