+++ src/parsers/i3d/loadCompressedGeometryData.ts
