+++ src/parsers/i3d/loadSectorMetadata.ts
