+++ src/parsers/i3d/loadUncompressedValues.ts
