+++ src/parsers/i3d/sharedFileParserTypes.ts
@@ -1,4 +1,4 @@
-import FibonacciDecoder from './FibonacciDecoder';
+import FibonacciDecoder from '../FibonacciDecoder';
 import { BYTES_PER_NODE_ID, geometryNameType } from './parserParameters';
 import * as THREE from 'three';
